+++ packages/whatsapp_bot_flutter_web/CHANGELOG.md
@@ -1,4 +1,8 @@
-## 0.0.2
+## 0.0.4
+
+- Bump whatsapp_bot_platform_interface to 2.0.2 and added `wppLibraryUrl` and `wppLibraryVersion` in `connect` method
+
+## 0.0.3
 
 - Update whatsapp_bot_platform_interface to 2.0.1